+++ test/interactive/interactive_base.hpp
@@ -33,6 +33,7 @@ using namespace arsh;
 #define CTRL_N "\x0E"
 #define CTRL_P "\x10"
 #define CTRL_R "\x12"
+#define CTRL_S "\x13"
 #define CTRL_T "\x14"
 #define CTRL_U "\x15"
 #define CTRL_V "\x16"